修改binsplit代码，添加make文件

+++ src/tools/bop.h
@@ -1,6 +1,6 @@
 #ifndef __TOOLS_H_
 #define __TOOLS_H_
 
-long strnum2long(char * str);
+unsigned long long strnum2long(char * str);
 
 #endif
